+++ src/optint_native.mli
@@ -1,0 +1,3 @@
+type t [@@immediate]
+
+include Integer_interface.S with type t := t